refactor : controller
컨트롤러 3개 빼고 삭제.
구현안 안되어 있음.
