--- src/test/java/team3/OneSubscribe/controller/signup/SignupControllerTest.java
@@ -1,10 +1,0 @@
-package team3.OneSubscribe.controller.signup;
-
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-@WebMvcTest
-class SignupControllerTest {
-
-}
