+++ src/test/java/team3/OneSubscribe/service/MemberServiceTest.java
@@ -30,7 +30,6 @@ class MemberServiceTest {
         Member member = new Member();
         member.setNickName("kim");
         member.setLoginId("testId");
-
         //When
         Long saveId = memberService.save(member);
